turned main.js into a module and separated each chart function to separate js files
